% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp025q01.wgt.R
\docType{data}
\name{cp025q01.wgt}
\alias{cp025q01.wgt}
\title{Treated data for PISA 2012, CP025, Q01 (selected countries)}
\format{
A data frame with 1456 rows and 15 variables.
}
\source{
\url{https://www.frontiersin.org/articles/10.3389/fpsyg.2019.02461/full}
}
\usage{
cp025q01.wgt
}
\description{
A dataset containing the weighted features generated from 2012 PISA Climate Control CP025Q01 task
}
\keyword{datasets}
