% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPartition.R
\name{DataPartition}
\alias{DataPartition}
\title{Data Partition}
\usage{
DataPartition(data = NULL, outcome = NULL, proportion = 0.7, seed = 2022)
}
\arguments{
\item{data}{A \code{data.frame} that contains the study’s features and the outcome variable.}

\item{outcome}{A character string with the name of the outcome variable from the data.}

\item{proportion}{A numeric value for the proportion of data to be put into model training. Default is set to 0.7.}

\item{seed}{A numeric value for set.seed. It is set to be 2022 by default.}
}
\value{
This function returns a list with training and testing data sets using a stratified selection by the outcome variable as performed by the createDataPartition function from the caret package.
}
\description{
Data Partition
}
\examples{
dp <- DataPartition(data = cp025q01.wgt, outcome = "outcome")
}
