% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DtResult.R
\name{DtResult}
\alias{DtResult}
\title{Decision Tree Result in Text View and Plot}
\usage{
DtResult(ModelObject)
}
\arguments{
\item{ModelObject}{A fitted model object from TreeModels() or TreeModelsAllSteps() functions.}
}
\value{
This function returns the structure of the decision tree final model as a text view, and a plot of the rpart model object as displayed by the rpart.plot package.
}
\description{
Decision Tree Result in Text View and Plot
}
\examples{
\donttest{
colnames(training)[14] <- "perf"
ensemblist <- TreeModels(traindata = training,
methodlist = "dt",checkprogress = TRUE)

DtResult(ensemblist$ModelObject)
}
}
