% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.R
\docType{data}
\name{testing}
\alias{testing}
\title{PISA 2012, CP025, Q01 (selected countries) Testing Data Set}
\format{
A data frame with 436 rows and 14 variables.
}
\source{
\url{https://www.frontiersin.org/articles/10.3389/fpsyg.2019.02461/full}
}
\usage{
testing
}
\description{
A testing set partitioned from the cp025q01.wgt dataset with 30% observations
}
\keyword{datasets}
