% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Training.R
\docType{data}
\name{training}
\alias{training}
\title{PISA 2012, CP025, Q01 (selected countries) Training Data Set}
\format{
A data frame with 1020 rows and 14 variables.
}
\source{
\url{https://www.frontiersin.org/articles/10.3389/fpsyg.2019.02461/full}
}
\usage{
training
}
\description{
A training set partitioned from the cp025q01.wgt dataset with 70% observations
}
\keyword{datasets}
